backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0